Split content_rewriting.py into sub modules.
No functional changes.
