+++ src/warc2zim/content_rewriting/__init__.py
@@ -1,0 +1,2 @@
+from warc2zim.content_rewriting.html import HtmlRewriter
+from warc2zim.content_rewriting.css import CSSRewriter
